Convert to typescript the msg slash commands files
